LayoutTests/media/modern-media-controls/scrubber-support/scrubber-support-seek-back-expected.txt

Scrubbing backwards in a video file with one keyframe should not seek to time zero.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS scrubber.getBoundingClientRect().width became different from 0
PASS media.currentTime < originalTime is true
PASS media.currentTime became different from 0
PASS successfullyParsed is true

TEST COMPLETE
